+++ src/main/kotlin/no/ssb/metadata/vardef/integrations/dapla/services/DaplaTeamService.kt
@@ -4,6 +4,7 @@ import io.micronaut.context.annotation.Prototype
 import io.micronaut.core.annotation.Introspected
 import no.ssb.metadata.vardef.integrations.dapla.models.Group
 import no.ssb.metadata.vardef.integrations.dapla.models.Team
+import no.ssb.metadata.vardef.models.Owner
 
 /**
  * Interface representing a service for retrieving and validating teams and groups.
@@ -23,4 +24,21 @@ interface DaplaTeamService {
     fun isValidGroup(groupName: String): Boolean
 
     fun getGroup(groupName: String): Group?
+
+    companion object {
+        private const val DEVELOPERS_SUFFIX = "developers"
+
+        /**
+         * Is the given group the developers group
+         *
+         * @param group the group in question
+         * @return `true` if it's the developers group
+         */
+        fun isDevelopers(group: String): Boolean = group.matches(Regex("^.+-developers$"))
+
+        fun containsDevelopersGroup(owner: Owner): Boolean =
+            owner.groups.any {
+                it == "${owner.team}-$DEVELOPERS_SUFFIX"
+            }
+    }
 }
